# Simplify ranges representation