--- a/core/trino-spi/src/test/java/io/trino/spi/predicate/TestRange.java
+++ b/core/trino-spi/src/test/java/io/trino/spi/predicate/TestRange.java
@@ -13,15 +13,6 @@
  */
 package io.trino.spi.predicate;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.module.SimpleModule;
-import io.airlift.json.ObjectMapperProvider;
-import io.trino.spi.block.Block;
-import io.trino.spi.block.TestingBlockEncodingSerde;
-import io.trino.spi.block.TestingBlockJsonSerde;
-import io.trino.spi.type.TestingTypeDeserializer;
-import io.trino.spi.type.TestingTypeManager;
-import io.trino.spi.type.Type;
 import org.testng.annotations.Test;
 
 import static io.airlift.slice.Slices.utf8Slice;
@@ -263,37 +254,5 @@ public class TestRange
         assertThatThrownBy(() -> range1To3Exclusive.intersect(range3To10))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("Cannot intersect non-overlapping ranges");
-    }
-
-    @Test
-    public void testJsonSerialization()
-            throws Exception
-    {
-        TestingTypeManager typeManager = new TestingTypeManager();
-        TestingBlockEncodingSerde blockEncodingSerde = new TestingBlockEncodingSerde();
-
-        ObjectMapper mapper = new ObjectMapperProvider().get()
-                .registerModule(new SimpleModule()
-                        .addDeserializer(Type.class, new TestingTypeDeserializer(typeManager))
-                        .addSerializer(Block.class, new TestingBlockJsonSerde.Serializer(blockEncodingSerde))
-                        .addDeserializer(Block.class, new TestingBlockJsonSerde.Deserializer(blockEncodingSerde)));
-
-        Range range = Range.all(BIGINT);
-        assertEquals(range, mapper.readValue(mapper.writeValueAsString(range), Range.class));
-
-        range = Range.equal(DOUBLE, 0.123);
-        assertEquals(range, mapper.readValue(mapper.writeValueAsString(range), Range.class));
-
-        range = Range.greaterThan(BIGINT, 0L);
-        assertEquals(range, mapper.readValue(mapper.writeValueAsString(range), Range.class));
-
-        range = Range.greaterThanOrEqual(VARCHAR, utf8Slice("abc"));
-        assertEquals(range, mapper.readValue(mapper.writeValueAsString(range), Range.class));
-
-        range = Range.lessThan(BIGINT, Long.MAX_VALUE);
-        assertEquals(range, mapper.readValue(mapper.writeValueAsString(range), Range.class));
-
-        range = Range.lessThanOrEqual(DOUBLE, Double.MAX_VALUE);
-        assertEquals(range, mapper.readValue(mapper.writeValueAsString(range), Range.class));
     }
 }