… core/trino-main/src/main/java/io/trino/server/RangeJsonSerde.java
… core/trino-main/src/test/java/io/trino/server/TestRangeJsonSerde.java
… core/trino-spi/src/main/java/io/trino/spi/predicate/SortedRangeSet.java