This changes `SortedRangeSet` representation from a `TreeSet<Range>` to a sorted list.


### before the change

```
Benchmark                       (stage)  Mode  Cnt     Score    Error  Units
BenchmarkPlanner.planQueries  optimized  avgt   20  1814.919 ± 81.205  ms/op
BenchmarkPlanner.planQueries    created  avgt   20    53.416 ±  5.858  ms/op

Benchmark                                 Mode  Cnt     Score     Error  Units
BenchmarkSortedRangeSet.benchmarkBuilder  avgt   10     2.427 ±   0.455  ms/op
BenchmarkSortedRangeSet.complementLarge   avgt   10  1746.084 ± 174.620  ms/op
BenchmarkSortedRangeSet.complementSmall   avgt   10  1484.053 ± 228.682  ms/op
BenchmarkSortedRangeSet.equalsLarge       avgt   10     0.190 ±   0.032  ms/op
BenchmarkSortedRangeSet.equalsSmall       avgt   10    39.038 ±   7.110  ms/op
BenchmarkSortedRangeSet.overlapsLarge     avgt   10    10.574 ±   2.585  ms/op
BenchmarkSortedRangeSet.overlapsSmall     avgt   10    90.876 ±  43.161  ms/op
BenchmarkSortedRangeSet.unionLarge        avgt   10   939.861 ± 188.176  ms/op
BenchmarkSortedRangeSet.unionSmall        avgt   10   298.672 ±  93.886  ms/op
```

### after the change

```
Benchmark                       (stage)  Mode  Cnt     Score    Error  Units
BenchmarkPlanner.planQueries  optimized  avgt   20  1766.144 ± 82.016  ms/op
BenchmarkPlanner.planQueries    created  avgt   20    54.039 ±  5.465  ms/op

Benchmark                                 Mode  Cnt    Score    Error  Units
BenchmarkSortedRangeSet.benchmarkBuilder  avgt   10    1.330 ±  0.795  ms/op
BenchmarkSortedRangeSet.complementLarge   avgt   10   47.965 ± 10.187  ms/op
BenchmarkSortedRangeSet.complementSmall   avgt   10  243.990 ± 45.498  ms/op
BenchmarkSortedRangeSet.equalsLarge       avgt   10    0.063 ±  0.006  ms/op
BenchmarkSortedRangeSet.equalsSmall       avgt   10   16.109 ±  3.119  ms/op
BenchmarkSortedRangeSet.overlapsLarge     avgt   10   11.015 ±  1.008  ms/op
BenchmarkSortedRangeSet.overlapsSmall     avgt   10  120.238 ± 16.833  ms/op
BenchmarkSortedRangeSet.unionLarge        avgt   10  306.094 ± 40.502  ms/op
BenchmarkSortedRangeSet.unionSmall        avgt   10  433.342 ± 75.544  ms/op
```