@findepi 2021-01-08
> Benchmark errors are huge. Can you try to decrease them somehow? Especially the planner benchmark

@skrzypo987 This is kind of "integration benchmark" without focus. I quote it before it showed huge regressions for the interim version of the code, so it is not insignificant either.
Do you have suggestions?