@martint 2021-01-08
You can run more forks and measurement iterations to get better bounds.

Also, I’d recommend plugging the results into  https://jmh.morethan.io/ to visualize and make the comparison easier.