@skrzypo987 2021-01-08
Adding `-Xbatch` VM arg will sometime reduce the difference between forks. Worked well in `BenchmarkDecimalOperators`.