@findepi 2021-01-08
With 
```
@Fork(value = 3, jvmArgsAppend = {"-Xbatch", "-server"})
```

the result are different, worse (both before & after) and the variability did not decrease (increased actually)

```
before


Benchmark                       (stage)  Mode  Cnt     Score     Error  Units
BenchmarkPlanner.planQueries  optimized  avgt   60  2561.870 ± 293.128  ms/op
BenchmarkPlanner.planQueries    created  avgt   60   161.525 ±  98.176  ms/op

after

Benchmark                       (stage)  Mode  Cnt     Score     Error  Units
BenchmarkPlanner.planQueries  optimized  avgt   60  2304.540 ± 277.962  ms/op
BenchmarkPlanner.planQueries    created  avgt   60   166.869 ± 148.829  ms/op
```