@findepi 2021-01-09
i realized i didn't benchmark one of the important methods (`containsValue`).

```
before

Benchmark                                   Mode  Cnt     Score     Error  Units
BenchmarkSortedRangeSet.containsValueLarge  avgt   10  2174.679 ± 101.515  ms/op
BenchmarkSortedRangeSet.containsValueSmall  avgt   10  1819.460 ±  88.682  ms/op

after

Benchmark                                   Mode  Cnt     Score    Error  Units
BenchmarkSortedRangeSet.containsValueLarge  avgt   10  1082.838 ± 94.465  ms/op
BenchmarkSortedRangeSet.containsValueSmall  avgt   10   346.605 ± 52.309  ms/op
```

i still need to simplify the code though