@skrzypo987 2021-01-11
Strange. Just run it on my laptop:
```
Benchmark                       (stage)  Mode  Cnt     Score    Error  Units
BenchmarkPlanner.planQueries  optimized  avgt   20  1830,421 ± 39,354  ms/op
BenchmarkPlanner.planQueries    created  avgt   20    50,713 ±  2,028  ms/op
```
and the errors are far better.
But I think that gains are big enough to ignore the benchmark errors for now.